+++ extensions/linear/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Linear Changelog
 
+## [Use OAuth utils] - 2024-02-01
+
+- Use new OAuth utils
+
 ## [Fix projects search not working] - 2024-01-30
 
 - The projects search was not working in case there were no roadmaps available. This is now fixed.